CreateVisualテンプレート版追加 FWObjectに便利APIとしてLoadMeshを追加
その他warning消し

git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4678 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHIKActuator.cpp
@@ -105,7 +105,7 @@ void PHIKActuator::PrepareSolve(){
 			//(*dlog) << "--- v[cp:" << eef_n << "] ---" << std::endl;
 			//(*dlog) << eef_v << std::endl;
 
-			for (size_t k=0; k < (*eef)->ndof; ++k) {
+			for (int k=0; k < (*eef)->ndof; ++k) {
 
 				// ÉøÅAÉ¿
 				alpha[i] += ( (Mj[eef_n][k][i]/bias) * (Mj[eef_n][k][i]) );
